enhancement: Add ReferenceNodeNotFoundException.
When inserting a new node before or after another node,
if the reference node does not exist thorw an exception.

Also, refactors insert tests into a separate test.
